wrong parser for scale in config
fixes #19 that was origin reported to https://github.com/deepnest-io/Deepnest/issues/199

+++ main/deepnest.js
@@ -567,8 +567,8 @@
         config.timeRatio = n;
       }
 
-      if (c.scale && parseInt(c.scale) > 0) {
-        config.scale = parseInt(c.scale);
+      if (c.scale && parseFloat(c.scale) > 0) {
+        config.scale = parseFloat(c.scale);
       }
 
       SvgParser.config({
